+++ Pod/FireFlower.h
@@ -35,10 +35,10 @@ public:
 	bool getIsFlip();
 	bool getIsGreenMode();
 	bool getIsHalfSizeMode();
-	float getWidth();
-	float getHeight();
-	float getBoundsWidth();
-	float getBoundsHeight();
+	int getWidth();
+	int getHeight();
+	int getBoundsWidth();
+	int getBoundsHeight();
 	float getTopAnchor();
 	float getBottomAnchor();
 	float getLeftAnchor();
